Refactoring APIs by purpose. cat drug etc..
